--- src/main/kotlin/com/deepromeet/atcha/transit/infrastructure/client/TMapTransitClient.kt
@@ -1,16 +1,0 @@
-package com.deepromeet.atcha.transit.infrastructure.client
-
-import com.deepromeet.atcha.transit.infrastructure.client.request.TMapRouteRequest
-import com.deepromeet.atcha.transit.infrastructure.client.response.TMapRouteResponse
-import org.springframework.cloud.openfeign.FeignClient
-import org.springframework.web.bind.annotation.GetMapping
-
-@FeignClient(
-    name = "tmap",
-    url = "\${tmap.api.url}",
-    configuration = [TMapFeignConfig::class]
-)
-interface TMapTransitClient {
-    @GetMapping("/transit/routes")
-    fun getRoutes(request: TMapRouteRequest): TMapRouteResponse
-}
